test_researcher_migration.py
"""
Phase 2: Researcher Migration Validation Script

This script compares the Strands Researcher implementation with the
original OpenAI Agents implementation to ensure feature parity.

Tests:
1. Both implementations can be created successfully
2. Both can handle the same research queries
3. MCP tools are accessible in both
4. Response quality is comparable
"""

import asyncio
import sys
from contextlib import AsyncExitStack
from dotenv import load_dotenv

load_dotenv(override=True)


async def test_strands_researcher_creation():
    """Test 1: Create Strands Researcher agent"""
    print("\n=== Test 1: Strands Researcher Creation ===")
    
    try:
        from strands_researcher import get_strands_researcher
        
        researcher = await get_strands_researcher("TestTrader", "gpt-4o-mini")
        print(f"✓ Created Strands researcher: {researcher.name}")
        
        # Check if agent has tools
        if hasattr(researcher, '_tools') or hasattr(researcher, 'tools'):
            print(f"  ✓ Agent has tools configured")
        
        # Cleanup
        if hasattr(researcher, 'cleanup') and callable(researcher.cleanup):
            await researcher.cleanup()
        
        return True
    except Exception as e:
        print(f"✗ Failed to create Strands researcher: {e}")
        import traceback
        traceback.print_exc()
        return False


async def test_openai_agents_researcher_creation():
    """Test 2: Create OpenAI Agents Researcher agent"""
    print("\n=== Test 2: OpenAI Agents Researcher Creation ===")
    
    try:
        from traders import get_researcher
        from agents.mcp import MCPServerStdio
        from mcp_params import researcher_mcp_server_params
        
        # Create MCP servers
        async with AsyncExitStack() as stack:
            mcp_servers = [
                await stack.enter_async_context(
                    MCPServerStdio(params, client_session_timeout_seconds=120)
                )
                for params in researcher_mcp_server_params("TestTrader")
            ]
            
            researcher = await get_researcher(mcp_servers, "gpt-4o-mini")
            print(f"✓ Created OpenAI Agents researcher: {researcher.name}")
            return True
    except Exception as e:
        print(f"✗ Failed to create OpenAI Agents researcher: {e}")
        import traceback
        traceback.print_exc()
        return False


async def test_strands_researcher_invocation():
    """Test 3: Invoke Strands Researcher with simple query"""
    print("\n=== Test 3: Strands Researcher Invocation ===")
    
    try:
        from strands_researcher import get_strands_researcher
        
        researcher = await get_strands_researcher("TestTrader", "gpt-4o-mini")
        
        # Simple query that should work quickly
        query = "What is the current stock market sentiment? Keep your answer brief."
        
        print(f"Query: {query}")
        print("Invoking researcher... (this may take 30-60 seconds)")
        
        result = await researcher.invoke_async(query)
        
        print(f"✓ Researcher responded successfully")
        print(f"  Stop reason: {result.stop_reason}")
        
        # Try to extract response text
        if hasattr(result, 'message'):
            msg = result.message
            if isinstance(msg, dict) and 'content' in msg:
                content = msg['content']
                if isinstance(content, list) and len(content) > 0:
                    response_text = content[0].get('text', '')[:200]
                    print(f"  Response preview: {response_text}...")
        
        # Cleanup
        if hasattr(researcher, 'cleanup') and callable(researcher.cleanup):
            await researcher.cleanup()
        
        return True
    except Exception as e:
        print(f"✗ Strands researcher invocation failed: {e}")
        import traceback
        traceback.print_exc()
        return False


async def test_openai_agents_researcher_invocation():
    """Test 4: Invoke OpenAI Agents Researcher with simple query"""
    print("\n=== Test 4: OpenAI Agents Researcher Invocation ===")
    
    try:
        from traders import get_researcher
        from agents import Runner
        from agents.mcp import MCPServerStdio
        from mcp_params import researcher_mcp_server_params
        
        # Create MCP servers
        async with AsyncExitStack() as stack:
            mcp_servers = [
                await stack.enter_async_context(
                    MCPServerStdio(params, client_session_timeout_seconds=120)
                )
                for params in researcher_mcp_server_params("TestTrader")
            ]
            
            researcher = await get_researcher(mcp_servers, "gpt-4o-mini")
            
            # Simple query
            query = "What is the current stock market sentiment? Keep your answer brief."
            
            print(f"Query: {query}")
            print("Invoking researcher... (this may take 30-60 seconds)")
            
            await Runner.run(researcher, query, max_turns=10)
            
            print(f"✓ OpenAI Agents researcher responded successfully")
            
            return True
    except Exception as e:
        print(f"✗ OpenAI Agents researcher invocation failed: {e}")
        import traceback
        traceback.print_exc()
        return False


async def test_mcp_tool_availability():
    """Test 5: Verify MCP tools are available"""
    print("\n=== Test 5: MCP Tool Availability ===")
    
    try:
        from strands_researcher import get_strands_researcher
        
        researcher = await get_strands_researcher("TestTrader", "gpt-4o-mini")
        
        # Check if MCP tools are loaded
        # This is implementation-specific, so we just verify no errors
        print("✓ Strands researcher created with MCP configuration")
        print("  Expected tools: brave_web_search, fetch, memory operations")
        
        # Cleanup
        if hasattr(researcher, 'cleanup') and callable(researcher.cleanup):
            await researcher.cleanup()
        
        return True
    except Exception as e:
        print(f"✗ MCP tool check failed: {e}")
        return False


async def main():
    """Run all validation tests"""
    print("=" * 60)
    print("Researcher Migration - Phase 2 Validation")
    print("=" * 60)
    
    results = []
    
    # Test 1: Strands researcher creation
    results.append(await test_strands_researcher_creation())
    
    # Test 2: OpenAI Agents researcher creation
    results.append(await test_openai_agents_researcher_creation())
    
    # Test 3: Strands researcher invocation (may take time)
    print("\n⚠ Note: Invocation tests may take 30-60 seconds each")
    results.append(await test_strands_researcher_invocation())
    
    # Test 4: OpenAI Agents researcher invocation (may take time)
    results.append(await test_openai_agents_researcher_invocation())
    
    # Test 5: MCP tool availability
    results.append(await test_mcp_tool_availability())
    
    # Summary
    print("\n" + "=" * 60)
    print("Validation Summary")
    print("=" * 60)
    
    test_names = [
        "Strands Researcher Creation",
        "OpenAI Agents Researcher Creation",
        "Strands Researcher Invocation",
        "OpenAI Agents Invocation",
        "MCP Tool Availability",
    ]
    
    for i, (name, result) in enumerate(zip(test_names, results), 1):
        status = "✓ PASS" if result else "✗ FAIL"
        print(f"Test {i}: {name:<35} {status}")
    
    passed = sum(results)
    total = len(results)
    print(f"\nTotal: {passed}/{total} tests passed")
    
    # Determine overall success
    if all(results):
        print("\n✅ Phase 2 validation SUCCESSFUL!")
        print("Researcher agent migration complete!")
        print("Ready to proceed to Phase 3 (Nested Agent Pattern)")
        return True
    else:
        print("\n⚠ Phase 2 validation INCOMPLETE")
        print("Some tests failed - review above for details")
        return False


if __name__ == "__main__":
    try:
        success = asyncio.run(main())
        sys.exit(0 if success else 1)
    except KeyboardInterrupt:
        print("\n\nValidation interrupted by user")
        sys.exit(1)
    except Exception as e:
        print(f"\n\nUnexpected error during validation: {e}")
        import traceback
        traceback.print_exc()
        sys.exit(1)
