requirements.txt
gradio>=5.7.0
openai==1.85.0
openai-agents==0.0.17
strands-agents>=1.0.0
pypdf==6.1.0
python-dotenv==1.2.1
requests==2.32.5
huggingface-hub==0.32.4
sendgrid==6.12.3
pydantic==2.11.5
nest-asyncio==1.6.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
wikipedia>=1.4.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
langgraph-checkpoint-sqlite>=3.0.1
typing-extensions>=4.8.0
markdownify==1.1.0
markdown-it-py==3.0.0
markdown-pdf==1.10
pymupdf==1.26.7
polygon-api-client==1.14.6
plotly==6.1.2